backend/requirements.txt
quart==0.18.4
Werkzeug==2.2.2
mysql-connector==2.2.9
requests==2.31.0
openai==0.28.0
aiohttp==3.8.5
rect-intersection==0.10
tiktoken==0.5.1
quart-cors==0.7.0
quart-session==2.0.0
reportlab==4.0.6
filelock==3.13.1
redis==4.4.0
stripe==8.0.0
pypdf==4.0.1
python-pptx==0.6.23